feat: implement transaction management and holding services for portfolio stock operations

+++ stockportfolio/src/main/java/com/sunil/stockportfolio/entity/Holding.java
@@ -22,5 +22,5 @@ public class Holding {
     @ManyToOne
     private Stocks stock;
     private Integer quantity;
-    private Integer averagePrice;
+    private BigDecimal averagePrice;
 }
